Create update_prod.yml
This GitHub action is designed to update the production branch with the latest changes from the main branch. Here’s a detailed explanation of what’s happening in this workflow:

The workflow is triggered manually using the workflow_dispatch event. This means that you can run this workflow whenever you want by going to the “Actions” tab in your repository and clicking on the “Run workflow” button.

The update-production job is defined to run on an ubuntu-latest virtual machine.

The first step in this job is to check out your repository using the actions/checkout@v2 action. This action checks out your repository’s code at the specified reference, which in this case is main.

The second step in this job is to push changes from the main branch to the production branch using a simple git command: git push origin main:production. This command pushes all changes from the local copy of the main branch to the remote production branch.

After running this workflow, your production branch will be updated with all changes from your main branch.